Changed to long double

+++ ComplexNumber.cpp
@@ -3,10 +3,10 @@
 ComplexNumber ComplexNumber::dZ(1,0);
 const double ComplexNumber::ARBRITARY_SMALL_NUM(.001f);
 
-ComplexNumber::ComplexNumber(double _a, double _b): a(_a),b(_b) {
+ComplexNumber::ComplexNumber(long double _a, long double _b): a(_a),b(_b) {
 }
 
-int ComplexNumber::getNumEscapeSteps(const ComplexNumber& c) {
+short ComplexNumber::getNumEscapeSteps(const ComplexNumber& c) {
     dZ.a=1;
     dZ.b=0;
     z.a=0;
@@ -24,7 +24,7 @@ int ComplexNumber::getNumEscapeSteps(const ComplexNumber& c) {
     return -1;
 }
 
-double ComplexNumber::magnitudeSquared() {
+long double ComplexNumber::magnitudeSquared() {
     return a*a+b*b;
 }
